backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
python-dotenv==1.0.0
openai==0.28.0
psycopg2-binary==2.9.7
pytest==7.4.2 